Add files via upload
Added Chapter 5 to Chapter 3 and 4 because it covers evaluation metrics. I specifically wanted to highlight how to intepret precision, recall, f1 macro using the classification_report feature as the most important take away from chapter 5 with discussion questions to allow users to contemplate the implications of precision, recall, and f1 macro. chapter 6 and 7 will be covered in a separate file because this is where the book pivots to text processing.