Add files via upload
Added Chapter 5, which covers evaluation metrics such as precision, recall, and f1 macro. The most important take away from Chapter 5 is how to interpret a classification report. Provided an example of how to interpret a classification report with thoughtful discussion questions on the implications of precision, recall, and f1 macro.